refactor(store): extract PTY buffers and settings parsers out of store.ts
Move two store-independent concerns out of the 2056-line store.ts into
dedicated modules, shrinking it to ~1885 lines:

- src/pty/buffers.ts: per-agent + side-shell PTY ring buffers and live
  sinks. Maps are now private; expose pushAgentOutput/pushShellOutput
  (append + dispatch) plus the get/clear/register helpers. The duplicated
  buffer grow/trim logic folds into one appendToRing helper.
- src/storage/preferences.ts: settings-blob parsers (parseFeatures and its
  legacy migration, parseProviderFlags, parsePaneWidth,
  parseProviderPathOverrides), the pane-width/AGENT_BIN_PREFIX constants,
  and the ThemeMode/Density/WorkspaceView/SettingsSection/FeatureFlags
  types. Lives next to storage/settings.ts whose output it parses.

interruptedAgents stays in store.ts (chime-suppression logic, not PTY
buffering). store.ts re-exports the preference types since they're part of
its public state shape, so existing consumers are unchanged. NativeView and
TermPanel now import buffer helpers directly from pty/buffers; the
provider-overrides test imports from storage/preferences.

tsc clean; full vitest suite green (295 tests).
